+++ dash_generator/example/main.dart
@@ -1,8 +1,9 @@
 import 'bloc_example.dart';
 import 'provider.dart';
 
-class Main{
-  void anyMethod(){
+class Main {
+  void anyMethod() {
     final bloc = $Provider.of<BlocExample>();
+    $Provider.dispose<BlocExample>();
   }
 }