+++ dash_generator/CHANGELOG.md
@@ -1,3 +1,3 @@
-## [0.2.0] - February second 2019.
+## [0.4.0] - February second 2019.
 
-* Changed cache class name.
+* Changed the way of how blocs are disposed